base_scraper/requirements.txt
playwright
beautifulsoup4
httpx
python-dotenv
tldextract